titan-poss-poss-web-f78a2c9ef984/titan-poss-poss-web-f78a2c9ef984/libs/shared/product-group-mapping/data-access-product-group-mapping/src/lib/+state/product-group-mapping.selector.ts
import { createSelector } from '@ngrx/store';
import { selectProductGroupMappingState } from './product-group-mapping.reducer';

const selectProductGroups = createSelector(
  selectProductGroupMappingState,
  state => state.productGroups
);

const selectError = createSelector(
  selectProductGroupMappingState,
  state => state.error
);

const selectIsloading = createSelector(
  selectProductGroupMappingState,
  state => state.isLoading
);

export const ProductGroupMappingSelectors = {
  selectProductGroups,
  selectError,
  selectIsloading
};
